*PushNotificationsManager.dart* Replaced var token = await _firebaseMessaging.getToken() with var token = await _firebaseMessaging.getAPNSToken() because getToken crashes on launch. This is likely not be permanent fix, as I see getToken may crash on simulator but not on real device. There are anecdotes of people who have getToken working on simulator. I will need to investigate further. The two current leads are 1) to check the configuration for the app in firebase to see if we have all the correct permissions enabled for notifications in IoS, and 2) to check if notification for the app in xcode is correctly configured.
*BackgroundTracking.dart*
In registerOneOffTask, I changed 'tracking_task_$index' to trackingTask_$index'.

*AppDelegate.swift*
Removed manual processing of BGProcessingTask. It worked fine, just was not needed.

*Info.plist*
UIApplicationSupportsIndirectInputEvents key was automatically added by Xcode, may not be needed.
UIBackgroundModes key was updated with remote-notification. This was part of the fix started in PushNotificationsManager.dart to get notifications working.

*build.gradle*
Updated ext.kotlin_version from '1.6.10' to '1.9.0'.
This was done to fix a build error that was caused by the kotlin version being too low.

*pubspec.yaml*
Updated environment sdk from '>=2.15.0 <3.0.0' to '>=2.17.1 <3.0.0'
Updated workmanager from ^0.5.0 to ^0.5.2

*Project.pbxproj*
Most changes were done automatically. Primary significant change by me was moving the SystemCapabilities code block shown below from the bottom of the file to
the PBXProject section below ProvisioningStyle. This was dome as it caused the file to show as being corrupted in Xcode.

SystemCapabilities = {
	com.apple.BackgroundModes = {
		enabled = 1;
	};
};

Some of the changes done automatically may not be necessary, with the exception of the SystemCapabilities changed, they can be reverted if needed.

*Launch.json*
Added while debugging camera issues. Resolution was that you can't access or test camera functionality via the simulator with IoS. So this may no longer be needed.