Changes to compile iOS + permissions for Background tracking on iOS (#252)
* *PushNotificationsManager.dart*
Replaced var token = await _firebaseMessaging.getToken() with var token = await _firebaseMessaging.getAPNSToken() because getToken
crashes on launch. This is likely not be permanent fix, as I see getToken may crash on simulator but not on real device. There are anecdotes
of people who have getToken working on simulator. I will need to investigate further. The two current leads are 1) to check the configuration for the app in firebase to
see if we have all the correct permissions enabled for notifications in IoS, and 2) to check if notification for the app in xcode is correctly configured.

*BackgroundTracking.dart*
In registerOneOffTask, I changed 'tracking_task_$index' to trackingTask_$index'.

*AppDelegate.swift*
Removed manual processing of BGProcessingTask. It worked fine, just was not needed.

*Info.plist*
UIApplicationSupportsIndirectInputEvents key was automatically added by Xcode, may not be needed.
UIBackgroundModes key was updated with remote-notification. This was part of the fix started in PushNotificationsManager.dart to get notifications working.

*build.gradle*
Updated ext.kotlin_version from '1.6.10' to '1.9.0'.
This was done to fix a build error that was caused by the kotlin version being too low.

*pubspec.yaml*
Updated environment sdk from '>=2.15.0 <3.0.0' to '>=2.17.1 <3.0.0'
Updated workmanager from ^0.5.0 to ^0.5.2

*Project.pbxproj*
Most changes were done automatically. Primary significant change by me was moving the SystemCapabilities code block shown below from the bottom of the file to
the PBXProject section below ProvisioningStyle. This was dome as it caused the file to show as being corrupted in Xcode.

SystemCapabilities = {
	com.apple.BackgroundModes = {
		enabled = 1;
	};
};

Some of the changes done automatically may not be necessary, with the exception of the SystemCapabilities changed, they can be reverted if needed.

*Launch.json*
Added while debugging camera issues. Resolution was that you can't access or test camera functionality via the simulator with IoS. So this may no longer be needed.

WIP

-- pubspec.yaml --
Changes below done to address the error of firebasemessaging get token crashing in simulator
Updated firebase_auth from 4.3.1 to 5.4.0
Updated firebase_messaging from 14.3.0 to 15.2.0
Updated Firebase_core from 2.8.0 to 3.10.0

Changes below done as part of flutter (updated flutter to 3.27.1)
Updated file_picker from ^5.2.10 to 8.0.3
Updated flutter_material_pickers from ^3.3.0 to 3.7.0
Updated language_picker from ^0.4.3 to 0.4.5
Updated flutter_html from ^3.0.0-alpha.6 to 3.0.0-beta.2

-- ios/Podfile --
Changed platform ios version from 12 to 16 to stay in line with apple app store requirements

-- AppDelegate.swift --
Modified to be in line with flutter v3.27.1
@UIApplicationMain no longer used, replaced with @main

-- notification_pages/notifications_page.dart --
Changes to flutter_html from ^3.0.0-alpha.6 to 3.0.0-beta.2 necessitated changes to the code

onLinkTap procedure updated to be inline with new flutter_html syntax
Location of style.dart page modified to be inline with new flutter_html file structure

-- settings_pages/settings_page.dart --
Changes to language_picker from ^0.4.3 to 0.4.5 necessitated changes to the code

language class now takes an additional parameter, NativeName, which is required

-- PushNotificationsManager.dart --
Gets getAPNSToken if the platform is iOS

temporary changed made to dev.json
drawer_and_header.dart changed due to error

changes in jp office after pulling this branch

Added ellipsis to avoid UUID number spilling over

* -- pubspect.yaml --
upgraded url_launcher from 6.2.4 to 6.32=.1
Added url_launcher_ios as direct dependency

-- static analysis.yml --
updated flutter version in file to 3.27.1

-- drawer_and_header.dart --
Commented out code on lines 58-65 work on physocal machine but not on simulator. Need to investigate further

* Added Info.plist translations

* Use ios/Runner folder instead of ios/Runner/Resources for iOS translations

* Rename webdev2 to webdev

* Localized language files generated by Xcode.

Re automatization, I recommend we first generate the file with Xcode, and then we can automatically update that generated file with whatever we want so long as we maintain the file name and folder structure.

* Localize all languages

* Added messages for notifications and tracking

* Remove trailing number on strings

* WorkManager registrys included in the AppDelegate
Workmanager package in pubspec changed to pull directly from a specific branch in git

ios build files generated due to manual changes made to the project.pbxproj file

* Revert some linters to make reviewing diff easier

* more lint

* more linter on drawer

* remove lint to be able to see diff

* add vm:entry-point + revert 2 changes

* iOS 13+ BGTaskSchedulerPermittedIdentifiers

* add debug case

* Upgrade android minSdk to 23

* try ios build

* workmanager: use specific branch

* try cancelAll() for iOS

* Revert "try ios build"

This reverts commit eca853c6add90f58eb5f488e07d14db855b5aca7.

* Target and Progect files auto generated because of manual entries to project file.

Podfile updated to enabled neceessary mocation permissions

UserManager updated to first request location permission when in use, then once granted requests location always on permission.

* solve merge conflict of pubspec.lock

* Update .gitignore

* delete build files

* remove more logs

* lint

* remove url_launcher_ios

* Update pubspec.yaml

* refactor to remove nesting

* remove debugging stuff on main

* remove unused import

---------

Co-authored-by: Otis Johnson <otis.johnson@upf.edu>
Co-authored-by: Enric Pou <enricpou@gmail.com>
Co-authored-by: Enric Pou <e.pou@ceab.csic.es>